Limit the length of providen name
In some cases GRIT may need to use the providen name as a sort of
namespace and add specific suffixes (when clarify on naming is really
important or when the template may create multiple resources of the same
type).

At the same time different resources in different cloud providers
sometimes have length restrictions.

We've chosen an arbitrary length of 12 to be the maximum for the name.
That should give us some margin for additional suffixes when needed.