Implement first iteration of the operator (internal) module
Current approach:
- use a helper that `updateManifest.sh` which download the operator
  manifests and restructures them, so that terraform can pick it up
  right away
- use `kubectl` terraform provider to deploy those resources

All in all, this seems to be the most reliable & straight-froward way I
could come up with.

Some questions you might have:

Why not just use the `kubernetes` provider, specifically the
`kubernetes_manifest`?

The resource type `kubernetes_manifest` needs the APIServer to be up and
running when doing a plan. Also, I have seen issues with configuring the
provider outside of the `operator` module, with inputs from the
`cluster` module, and setting up dependencies between the `cluster`
module and the `operator` module.
Also ordering resources (have the namespace deployed before other
things) is quite tricky.
Thus the switch to the `kubectl` provider.

Why this `yq` massaging of the data? Can't we use the
`kubectl_file_documents` data source and then a `for_each` in the
`kubectl_manifest` resource?

We want to configure the provider outside of the `operator` module.
Even though the `kubectl` provider does not need any configuration to use
the `kubectl_file_documents` data resource, terraform still bails out,
because it says it cannot calculate the for_each, because it can't use
the data resource until it has deployed the `cluster` and can get the
creds/coords thereof.

Why not just download the manifest and wrap it in a `v1/List`?

The `kubectl` provider does not support `v1/List`s. For the `kubernetes`
provider we'd need to use the `kubernetes_manifest` resources and have
all the same issues again.