Introduce example to deploy the operator on GKE
This uses the test / internal modules (for now), to bring up a GKE
cluster and deploy the operator onto.
Providers are defined in the root module and (implicitely) passed to all
the submodules, which allows users to configure their providers as
needed, but also to set up dependencies between modules.