Introduce the k8s/runner module
Using the operator, we can now deploy a runner and wire that together
with the gitlab module, to register it to the GitLab instance.

In the default configuration the operator only picks up `Runner`s in a
specific namespace, so we also need to wire that from the operator
module to the runner module.