Allow configuration of operator version
We support multiple versions of the operator now. We pull all the
released versions from the operator upstream into this repo, so that we
don't have to do any additional http calls at runtime.
However, we also provide an escape hatch, where users can provide a path
to a file holding the operator deployment manifests.