[AWS] Allow to choose the gitlab-runner package repository
Until now the AWS module was enforcing the usage of
https://packages.gitlab.com/runner/gitlab-runner repository, limiting
the user to only stable releases.

That's okay from the production setup perspective, but it limits
administrators capabilities of testing the changes from GitLab Runner
`main` branch before their are released with the tagged version.

With this change the repository is now configurable. While it still
set to `gitlab-runner` (the stable one) by default - so there are no
changes for the existing users of GRIT - it can be overwritten to
`unstable` to use the `pre` versions from
https://packages.gitlab.com/runner/unstable.