Lock registered project runners
This commit makes the registered project runners locked by default.

A project runner may be shared between different projects on the same
GitLab instance. When it's done, being a single `ci_runners` entity with
the same ID and token, it's available for multiple projects where it was
additionally assigned (even across top-level groups). There is a
specific way how it can be done, which I'll not describe here as it's
documented elsewhere.

What's important is that in most cases we don't want to allow a project
runner to be shared with other projects. And that's what the `locked`
flag in `ci_runners` entities is for. When set to `true`, runner is
locked to currently attached projects and new project can't be added.

Long time ago we've made `locked` to be set to `true` by default when
registering it through `gitlab-runner register ...` command. Today, with
the new registration flow, I think we have it default in UI. And maybe
in API.

But not in GRIT.

GRIT uses GitLab Terraform provider, this one has the `locked` attribute
for the `gitlab_user_runner` resource. And this one has a defauilt
value, which seems to be `false`. It means that when we register a new
runner through terraform, it's by default not locked.

For instance and group runners that doesn't matter, as `locked` flag is
unused in their cases. But for `project` runners it means, that this
project is by default available for sharing. Which shouldn't happen.

Runner may be always unlocked and locked back through UI/API, so we
should follow the good practices we've introduced in the past and set
the `locked` flag to true. If needed - runner owner will be able to
unlock it and share the runner with other projects.

And this is what this commit does :)