Improve e2e:state-clean job
This commit brings two changes:

1. Moves e2e:state-clean script definition to a dedicated script file.
   This makes both the YAML definition simpler and the script itself
   easier to maintain.

2. Updates the script so that it doesn't fail on the first locked state
   that is applicable for cleanup.

There may be cases where a state will be left in a locked state, like
when a failure happened in the middle of Terraform execution. Such state
can't be deleted until the existing lock is removed.

Current implementtion of `e2e:state-clean` would try to remove it anyway,
failing and therefore exiting the job before it would even approach to
remove the other states. If the first listed and applicable for deletion
state is locked, `e2e:state-clean` would keep failing for all pipelines.

This commit excludes such states from the removal loop, so that we don't
fail the job and we try to remove as much of stale states as possible in
each pipeline.

States older than 24 hours that are currently locked most probably need
a manual removal anyway. Therefore `e2e:state-clean` will now list them,
so we know what to look for.