Improve e2e:state-clean job
Follow-up for
https://gitlab.com/gitlab-org/ci-cd/runner-tools/grit/-/merge_requests/250+

Previously we've handled skipping cleanup of states that are left in a
locked state. Automatic cleanup will not be able to handle it and they
will require manual removal.

But we're still wait with states that were - for some reason - partially
removed during normal execution. And now, despite that the states are
not locked, they can't be easily removed. For example: a network is in
use by an instance. Instance resource was already removed from the
Terraform state, but it wasn't physically removed from the cloud.
Running `scripts/e2e-destroy.sh` Terraform fails, as it tries to remove
just the network and the cloud provider responds that it's "in use". But
Terraform has no data to handle that and it fails.

This commit catches such failures in a clean way, not failing the job,
and prints a list of all states that were failed on destroy.

It also marks the `e2e:state-clean` job as allowed to fail. We may get
future failures that we haven't anticipated yet. As this job was added
as a form of automated cleanup of old states, it shouldn't block
pipelines and merge requests from being merged.