Improve deployer ssh connection resilience
If SSH connection is unavailable immediately - in most cases it will be
when calling `wait-helathy` after terraform execution was finished but
before the runner manager instance was created and became available
through network - SSH based commands (`wait-healthy`, `shutdown`,
`wait-terminated`) will fail immediatelly.

This is not the ideal behavior, especially in the `wait-healthy` case.
We have the `--wrapper-connection-timeout` flag and the desire was that
it should make the command execution await until it successfully
connects to the wrapper or fail, but only after the timeout is exceeded.
That should cover also the SSH connection, not just the tunneled gRPC
connection.

This commit brings the needed changes.